+++ test/test_evolve.jl
@@ -62,7 +62,7 @@
         @test length(snaps) == 6
     end
     @testset "Gradient echo sequence" begin
-        simulation = mr.Simulation(mr.Sequence(pulses=[mr.InstantRFPulse(flip_angle=90)], TR=2.8))
+        simulation = mr.Simulation(mr.Sequence(components=[mr.InstantRFPulse(flip_angle=90)], TR=2.8))
         snaps = mr.trajectory(zeros(3), simulation, 0:0.5:2.8)
         @test mr.orientation(snaps[1]) ≈ SA[0., 0., 1.]
         for snap in snaps[2:end]
@@ -85,7 +85,7 @@
         @test mr.get_time(snaps) == 2.8
     end
     @testset "Basic diffusion has no effect in constant fields" begin
-        sequence = mr.Sequence(pulses=[mr.InstantRFPulse(flip_angle=90)], TR=2.)
+        sequence = mr.Sequence(components=[mr.InstantRFPulse(flip_angle=90)], TR=2.)
         no_diff = mr.Simulation([sequence], R2=0.3)
         with_diff = mr.Simulation([sequence], diffusivity=1., R2=0.3)
         spin_no_diff = mr.evolve(mr.Spin(), no_diff).spins[1]
@@ -97,7 +97,7 @@
         @test abs(mr.longitudinal(spin_no_diff)) < Float(1e-6)
     end
     @testset "Basic diffusion run within sphere" begin
-        sequence = mr.Sequence(pulses=[mr.InstantRFPulse(flip_angle=90)], TR=2.)
+        sequence = mr.Sequence(components=[mr.InstantRFPulse(flip_angle=90)], TR=2.)
         sphere = mr.Sphere(1.)
         Random.seed!(12)
         diff = mr.Simulation([mr.Sequence(TR=20.)], diffusivity=2., geometry=sphere)
@@ -112,9 +112,9 @@
     end
     @testset "Run simulation with multiple sequences at once" begin
         sequences = [
-            mr.Sequence(pulses=[mr.InstantRFPulse(flip_angle=0), mr.Readout(2.)], TR=3.),
-            mr.Sequence(pulses=[mr.InstantRFPulse(flip_angle=90), mr.Readout(2.)], TR=3.),
-            mr.Sequence(pulses=[mr.InstantRFPulse(flip_angle=90), mr.Readout(1.)], TR=2.),
+            mr.Sequence(components=[mr.InstantRFPulse(flip_angle=0), mr.Readout(2.)], TR=3.),
+            mr.Sequence(components=[mr.InstantRFPulse(flip_angle=90), mr.Readout(2.)], TR=3.),
+            mr.Sequence(components=[mr.InstantRFPulse(flip_angle=90), mr.Readout(1.)], TR=2.),
         ]
         all_snaps = mr.Simulation(sequences, diffusivity=1., R2=1.)